chore: Migrate to webpack and typescript
