+++ .babelrc
@@ -1,0 +1,10 @@
+{
+  "presets": [
+      "@babel/env",
+      "@babel/typescript"
+  ],
+  "plugins": [
+      "@babel/proposal-class-properties",
+      "@babel/proposal-object-rest-spread"
+  ]
+}
