+++ .eslintrc
@@ -1,0 +1,60 @@
+{
+  "rules": {
+    "no-cond-assign": 1,
+    "no-console": 1,
+    "no-constant-condition": 1,
+    "no-control-regex": 1,
+    "no-debugger": 1,
+    "no-dupe-args": 1,
+    "no-dupe-keys": 1,
+    "no-duplicate-case": 1,
+    "no-empty-character-class": 1,
+    "no-empty": 1,
+    "no-ex-assign": 1,
+    "no-extra-boolean-cast": 1,
+    "no-extra-semi": 1,
+    "no-func-assign": 1,
+    "no-inner-declarations": 1,
+    "no-invalid-regexp": 1,
+    "no-irregular-whitespace": 1,
+    "no-negated-in-lhs": 1,
+    "no-obj-calls": 1,
+    "no-regex-spaces": 1,
+    "no-sparse-arrays": 1,
+    "no-unreachable": 1,
+    "use-isnan": 1,
+    "valid-typeof": 1,
+    "no-alert": 1,
+    "no-else-return": 1,
+    "no-eval": 1,
+    "no-extra-bind": 1,
+    "no-fallthrough": 1,
+    "no-octal": 1,
+    "no-proto": 1,
+    "no-redeclare": [1, {"builtinGlobals": true}],
+    "no-useless-call": 1,
+    "no-delete-var": 1,
+    "no-undef-init": 1,
+    "no-undef": 1,
+    "no-unused-vars": 1,
+    "camelcase": 1,
+    "eol-last": 1,
+    "indent": [1, 2, {"SwitchCase": 1}],
+    "no-array-constructor": 1,
+    "no-continue": 1,
+    "no-lonely-if": 1,
+    "no-mixed-spaces-and-tabs": 1,
+    "no-new-object": 1,
+    "no-const-assign": 1,
+    "prefer-const": 1,
+    "prefer-spread": 1
+  },
+  "env": {
+    "es6": true,
+    "browser": true
+  },
+  "extends": "eslint:recommended",
+  "parserOptions": {
+    "sourceType": "module"
+  }
+}
