+++ .gitignore
@@ -1,1 +1,6 @@
 .vscode/
+node_modules/
+.env.local
+.env.development.local
+.env.test.local
+.env.production.local
