--- index.html
@@ -1,12 +1,0 @@
-<!-- Minimal html is placed here. May need more meta / style later. -->
-<!DOCTYPE html>
-<html>
-<head>
-  <title>Steel of Stalin</title>
-  <meta name="viewport" content="width=device-width, user-scalable=no, minimum-scale=1.0, maximum-scale=1.0">
-  <meta charset="utf-8">
-</head>
-<body>
-  <script src="./src/main.js" type="module"></script>
-</body>
-</html>
