+++ package.json
@@ -1,0 +1,57 @@
+{
+  "name": "comp4451",
+  "version": "1.0.0",
+  "description": "This is the repository for the game project in COMP4451. More Details to be added...",
+  "main": "index.js",
+  "scripts": {
+    "start": "webpack-dev-server"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/hkleungai/COMP4451.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/hkleungai/COMP4451/issues"
+  },
+  "homepage": "https://github.com/hkleungai/COMP4451#readme",
+  "devDependencies": {
+    "@babel/cli": "^7.2.3",
+    "@babel/core": "^7.8.7",
+    "@babel/plugin-proposal-class-properties": "^7.4.0",
+    "@babel/plugin-proposal-object-rest-spread": "^7.4.0",
+    "@babel/preset-env": "^7.8.7",
+    "@babel/preset-typescript": "^7.3.3",
+    "@types/node": "7.0.0",
+    "@types/three": "^0.125.3",
+    "autoprefixer": "^9.7.4",
+    "awesome-typescript-loader": "^5.2.1",
+    "babel-loader": "^8.0.6",
+    "copyfiles": "^2.2.0",
+    "css-loader": "^2.1.1",
+    "dotenv-flow": "^3.2.0",
+    "eslint": "^5.16.0",
+    "html-webpack-plugin": "^3.2.0",
+    "mini-css-extract-plugin": "^0.5.0",
+    "node-sass": "^4.13.1",
+    "npm-run-all": "^4.1.5",
+    "optimize-css-assets-webpack-plugin": "^5.0.3",
+    "postcss-loader": "^3.0.0",
+    "rimraf": "^2.7.1",
+    "sass-loader": "^7.3.1",
+    "source-map-loader": "0.2.3",
+    "style-loader": "^0.23.1",
+    "tslint": "5.17.0",
+    "tslint-loader": "3.6.0",
+    "typescript": "3.5.1",
+    "uglifyjs-webpack-plugin": "^2.2.0",
+    "webpack": "^4.42.0",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-server": "^3.10.3"
+  },
+  "dependencies": {
+    "es6-promise": "^4.2.8",
+    "three": "^0.126.0"
+  }
+}
