+++ src/assets/tiles/desert.png
