+++ src/assets/tiles/metropolis_blue.png
