+++ src/index.html
@@ -1,0 +1,10 @@
+<!-- Minimal html is placed here. May need more meta / style later. -->
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>Steel of Stalin</title>
+    <meta name="viewport" content="width=device-width, user-scalable=no, minimum-scale=1.0, maximum-scale=1.0">
+    <meta charset="utf-8">
+  </head>
+  <body><!-- Content of body tag is generated automatically --></body>
+</html>
