--- src/resources/index.js
@@ -1,3 +1,0 @@
-"use strict";
-
-export * from './textures.js';
