+++ src/utils/constants.ts
@@ -1,5 +1,3 @@
-"use strict";
-
 const THREE = 'https://cdn.jsdelivr.net/npm/three';
 const THREE_BUILD = `${THREE}/build/three.module.js`;
