+++ src/utils/helpers.ts
@@ -1,7 +1,5 @@
-"use strict";
-
 const randint = upper_bound => Math.floor(Math.random() * upper_bound);
-const range = upper_bound => [...Array(upper_bound).keys()];
+const range = (upper_bound: number) => [...Array(upper_bound).keys()];
 
 const sin_deg = deg => Math.sin(deg / 180 * Math.PI);
 const cos_deg = deg => Math.cos(deg / 180 * Math.PI);
