--- src/utils/index.js
@@ -1,4 +1,0 @@
-"use strict";
-
-export * from './constants.js';
-export * from './helpers.js';
